Countgamemainの作成
